requirement.txt
asgiref==3.8.1
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-admin-thumbnails==0.2.9
django-ckeditor==6.7.1
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-storages==1.14.4
django-taggit==6.1.0
idna==3.10
jmespath==1.0.1
pillow==11.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
